Move render strategy to module. Added create canvas to viewport initialization
