+++ example/index.html
@@ -10,8 +10,6 @@
       <input type="file" accept="image/png" id="input-image" />
       <button id="rect">Прямоугольники</button>
     </div>
-    <div class="canvas">
-      <canvas id="canvas" width="500" height="500"></canvas>
-    </div>
+    <div id="viewport"></div>
   </body>
 </html>
