+++ example/style.css
@@ -6,11 +6,10 @@ body {
   justify-content: center;
 }
 
-.canvas {
+#viewport {
   display: flex;
   justify-content: center;
-}
-
-#canvas {
+  width: 500px;
+  height: 500px;
   border: 2px solid black;
 }
