+++ src/index.ts
@@ -4,5 +4,10 @@ export type { Layer, LayerOptions, LayerEmitter } from './layer';
 export { CanvasLayer, LayerFactory } from './layer';
 export type { LayerListEmitter } from './layer-list';
 export { LayerList } from './layer-list';
-export type { RenderStrategy } from './viewport';
+export type {
+  RenderStrategy,
+  RenderStrategyConstructor,
+  RenderStrategyType,
+} from './render';
+export { CanvasRenderStrategy } from './render';
 export { Viewport } from './viewport';
