+++ src/render/canvas-render-strategy.ts
@@ -1,4 +1,4 @@
-import type { RenderStrategy } from '~/viewport';
+import type { RenderStrategy } from './interface';
 import { LayerList } from '~/layer-list';
 
 export default class CanvasRenderStrategy implements RenderStrategy {
@@ -17,11 +17,13 @@ export default class CanvasRenderStrategy implements RenderStrategy {
 
   render(): Promise<void> {
     return new Promise((resolve) => {
-      this.#clean();
-      for (const layer of this.#layers) {
-        this.#context.drawImage(layer.canvas, layer.offset.x, layer.offset.y);
-      }
-      resolve();
+      requestAnimationFrame(() => {
+        this.#clean();
+        for (const layer of this.#layers) {
+          this.#context.drawImage(layer.canvas, layer.offset.x, layer.offset.y);
+        }
+        resolve();
+      });
     });
   }
