+++ src/render/index.ts
@@ -1,0 +1,6 @@
+export type {
+  RenderStrategy,
+  RenderStrategyType,
+  RenderStrategyConstructor,
+} from './interface';
+export { default as CanvasRenderStrategy } from './canvas-render-strategy';
