+++ src/viewport/index.ts
@@ -1,6 +1,1 @@
 export { default as Viewport } from './viewport';
-export type {
-  RenderStrategy,
-  RenderStrategyType,
-  RenderStrategyConstructor,
-} from './interface';
